requirements.txt
pillow opencv-python
numpy
Pillow